SQL: RowDescription reports real Postgres type OIDs for computed columns
First external-gauge run (psycopg 3 suite + sqllogictest corpus) flagged
type-OID fidelity as the highest-leverage divergence: computed results
(CASE, array constructors/casts, integer arithmetic, bound parameters)
described as text (25), and smallint/real widened to integer/double
precision everywhere.

- typemap: first-class int2 (21) / float4 (700) tags — columns, casts,
  arrays (1005/1021), pg_type rows, information_schema spellings;
  SMALLSERIAL no longer describes as text.
- planner: CASE types from its result branches; array[...] and array
  casts report the array OID; integer arithmetic stays integer (matching
  _pg_div's truncating division); an unadorned decimal constant is
  numeric; sum(int2/int4) -> int8, sum(int8) -> numeric, avg(integer) ->
  numeric per Postgres' aggregate result types; CAST($1 AS SMALLINT)
  coerces its text-bound value numerically (un-breaks SQLAlchemy
  reflection joins on pg_attribute).
- pgextended: SELECT $1 describes with the parameter OID the client
  declared in Parse; binary result format + binary parameters now cover
  arrays (real ndim/hasnull/elemoid layout, both directions) — the
  correct array OIDs engage libpq clients' binary array parsers, which
  the text-bytes fallback would have fed garbage.

psycopg-gauge subset (test_connection/test_cursor/test_cursor_common/
types/test_{numeric,string,bool}): 409 -> 542 passed (42% -> 55%).

+++ tests/test_sql_compute.py
@@ -117,7 +117,8 @@ def test_computed_type_tags(storage, session):
         "SELECT price * qty AS total, upper(name) AS u, length(name) AS l FROM items WHERE _id = 1",
     )
     tags = {c.name: c.type_tag for c in res.columns}
-    assert tags == {"total": "numeric", "u": "text", "l": "int4"}
+    # price/qty are int8 columns; bigint * bigint stays bigint in Postgres.
+    assert tags == {"total": "int8", "u": "text", "l": "int4"}
 
 
 def test_computed_over_join(storage, session):
